chore: move Trial Information table in a dedicated Trial page tab [DET-5434 (#2372)
